# Use rsync to sync experiment directory for slurm