--- a/src/nemo_run/run/torchx_backend/schedulers/slurm.py
+++ b/src/nemo_run/run/torchx_backend/schedulers/slurm.py
@@ -50,7 +50,7 @@ from torchx.specs import (
 )
 from torchx.specs.api import is_terminal
 
-from nemo_run.config import NEMORUN_HOME, SCRIPTS_DIR, from_dict
+from nemo_run.config import NEMORUN_HOME, from_dict
 from nemo_run.core.execution.base import Executor
 from nemo_run.core.execution.slurm import SlurmBatchRequest, SlurmExecutor, SlurmJobDetails
 from nemo_run.core.tunnel.client import LocalTunnel, SSHTunnel, Tunnel
@@ -129,6 +129,16 @@ class SlurmTunnelScheduler(SchedulerMixin, SlurmScheduler):  # type: ignore
             launcher=executor.get_launcher(),
         )
 
+        # Write and copy sbatch script
+        sbatch_dir = executor.experiment_dir
+        path = os.path.join(sbatch_dir, f"{executor.job_name}_sbatch.sh")
+        script = req.materialize()
+
+        with open(path, "w") as f:
+            f.write(script)
+
+        executor.package(packager=executor.packager, job_name=Path(job_dir).name)
+
         return AppDryRunInfo(req, repr)
 
     def schedule(self, dryrun_info: AppDryRunInfo[SlurmBatchRequest]) -> str:  # type: ignore
@@ -145,32 +155,7 @@ class SlurmTunnelScheduler(SchedulerMixin, SlurmScheduler):  # type: ignore
         self._initialize_tunnel(tunnel)
         assert self.tunnel, f"Cannot initialize tunnel {tunnel}"
 
-        self.tunnel.connect()
-        self.tunnel.setup()
-
-        slurm_cfg.package(packager=slurm_cfg.packager, job_name=Path(job_dir).name)
-
-        # Write and copy sbatch script
-        sbatch_dir = slurm_cfg.experiment_dir or job_dir
-        path = os.path.join(sbatch_dir, f"{slurm_cfg.job_name}_sbatch.sh")
-        script = req.materialize()
-
-        with open(path, "w") as f:
-            f.write(script)
-
-        remote_scripts_dir = os.path.join(self.tunnel.job_dir, Path(job_dir).name, SCRIPTS_DIR)
-        for job in req.jobs:
-            local_script = os.path.join(job_dir, SCRIPTS_DIR, f"{job}.sh")
-            if os.path.isfile(local_script):
-                self.tunnel.run(f"mkdir -p {remote_scripts_dir}")
-                self.tunnel.put(
-                    local_script,
-                    os.path.join(remote_scripts_dir, Path(local_script).name),
-                )
-
         dst_path = os.path.join(self.tunnel.job_dir, f"{slurm_cfg.job_name}_sbatch.sh")
-        self.tunnel.put(local_path=path, remote_path=dst_path)
-
         # Run sbatch script
         req.cmd += [dst_path]
         job_id = self.tunnel.run(" ".join(req.cmd)).stdout.strip()