This also changes the order of shipping files via the tunnel. Previously, each task would ship its files sequentially. Moreover, since tasks were also launched sequentially, this essentially meant calling `sftp.put` on each file, which was extremely slow. It uses a modified version of https://github.com/fabric/patchwork/blob/master/patchwork/transfers.py to construct the rsync call.

This PR first prepares the entire experiment directory, then rsyncs it to the tunnel, then launches the job. 

With this PR: `rsync` will be a required tool, but most systems have it built in so should be fine for now.

Results:

Simple experiment with 5 job groups
```python
    inline_script = run.Script(
        inline="""
echo "Hello 1"
echo "Hello 2"
"""
    )

    with run.Experiment("slurm-demo", executor=executor, log_level="INFO") as exp:
        for i in range(5):
            exp.add([inline_script, inline_script], tail_logs=True, name=f"my-test-run-{i}")
        exp.run(detach=True, tail_logs=True, sequential=False)
```

Before this PR:
```
Benchmark 1: python test_slurm.py
  Time (mean ± σ):     62.132 s ±  0.912 s    [User: 4.972 s, System: 0.463 s]
  Range (min … max):   61.487 s … 62.776 s    2 runs
```

After this PR:
```
Benchmark 1: python test_slurm.py
  Time (mean ± σ):     22.131 s ±  1.450 s    [User: 5.655 s, System: 0.553 s]
  Range (min … max):   21.105 s … 23.156 s    2 runs
```

This will scale even better with more jobs as rsync is highly optimized.